Track API changes.

+++ src/model/tools/brush_imp.rs
@@ -215,7 +215,7 @@ pub fn draw_glyph(
         if let Some(glyph) = font.get_glyph(ch) {
             for y in 0..s.height {
                 for x in 0..s.width {
-                    if glyph.data[y] & (128 >> x) != 0 {
+                    if glyph.data[y as usize] & (128 >> x) != 0 {
                         painter.rect_filled(
                             Rect::from_min_size(
                                 Pos2::new(
