+++ src/model/tools/draw_ellipse_filled_imp.rs
@@ -114,8 +114,8 @@ impl Tool for DrawEllipseFilledTool {
         }
 
         let draw = move |rect: Rectangle| {
-            for y in 0..(rect.size.height as i32) {
-                for x in 0..(rect.size.width as i32) {
+            for y in 0..rect.size.height {
+                for x in 0..rect.size.width {
                     plot_point(
                         editor,
                         self,
