+++ src/model/tools/draw_rectangle_filled_imp.rs
@@ -108,8 +108,8 @@ impl Tool for DrawRectangleFilledTool {
         lines.add_rectangle(Rectangle::from_pt(start, cur));
 
         let draw = move |rect: Rectangle| {
-            for y in 0..(rect.size.height as i32) {
-                for x in 0..(rect.size.width as i32) {
+            for y in 0..rect.size.height {
+                for x in 0..rect.size.width {
                     plot_point(
                         editor,
                         self,
